added new methods and fixed previous

+++ transport/sanic/endpoints/users/user.py
@@ -5,11 +5,11 @@ from api.request import RequestPatchUserDto
 from api.response import ResponseGetUserDto
 
 from db.database import DBSession
-from db.exceptions import DBUserNotFoundException, DBIntegrityException, DBDataException
+from db.exceptions import DBUserNotFoundException, DBIntegrityException, DBDataException, DBUserDeletedException
 from db.queries import user as user_queries
 
 from transport.sanic.endpoints import BaseEndpoint
-from transport.sanic.exceptions import SanicUserNotFoundException, SanicDBException, SanicAuthException
+from transport.sanic.exceptions import SanicUserNotFoundException, SanicDBException, SanicUserDeletedException
 
 
 class UserEndpoint(BaseEndpoint):
@@ -19,21 +19,18 @@ class UserEndpoint(BaseEndpoint):
     ) -> BaseHTTPResponse:
 
         # проверяем, что пользователь посылает запрос от своего имени
-        # и что он не удален из базы (is_deleted != 1)
         if token.get('id') != user_id:
             return await self.make_response_json(status=403)
 
-        try:
-            user_queries.get_user(session=session, user_id=token['id']).is_deleted
-        except DBUserNotFoundException:
-            raise SanicUserNotFoundException('User not found')
-
         request_model = RequestPatchUserDto(body)
 
+        # проверяем, что пользователь есть в базе и не удален
         try:
             user = user_queries.patch_user(session, request_model, user_id)
         except DBUserNotFoundException:
             raise SanicUserNotFoundException('User not found')
+        except DBUserDeletedException:
+            raise SanicUserDeletedException('User deleted')
 
         try:
             session.commit_session()
@@ -45,11 +42,17 @@ class UserEndpoint(BaseEndpoint):
         return await self.make_response_json(status=200, body=response_model.dump())
 
     async def method_delete(
-            self, request: Request, body: dict, session: DBSession, user_id: int, *args, **kwargs
+            self, request: Request, body: dict, session: DBSession, user_id: int, token: dict, *args, **kwargs
     ) -> BaseHTTPResponse:
+
+        if token.get('id') != user_id:
+            return await self.make_response_json(status=403)
 
         try:
             user_queries.delete_user(session, user_id)
+        except DBUserDeletedException:
+            # TODO протестить
+            raise SanicUserDeletedException('User deleted')
         except DBUserNotFoundException:
             raise SanicUserNotFoundException('User not found')
